Move isLog2cpm() and isLogTransformed() to a new QuantitationTypeUtils helper
Ignore case when looking for log2cpm in the name.
